requirements.txt
mss>=9.0.1
Pillow>=10.0.0
numpy>=1.24.0
flask>=3.0.0
